Add an experimental sky model for realistic lighting without an HDR.
Right now it's not that useful, since we're not doing any direct sampling of lights.
Increasing the sample count with this in use will just result in a mega-noisy image.
But at least the sky looks pretty!

To enable this, you have to edit getBackground() in pathtrace.c to return the value from sky()