Add a neat interactive debugging script.
This script, when invoked with a path to a scene .json as an
argument, will compile, render that scene and open the output using
the 'open' command on macOS. I assume linux distros have a similar cmd,
so feel free to add support for those as well. Right now it assumes
you have a macOS-style 'open' in your $PATH.

Whenever you edit a source code file, or the scene you specified,
it will automatically recompile, run and show the resulting image.
Super fast debug/test cycles!