Add a simple alpha channel transparency implementation.
Currently it just uses the alpha channel as a probability to determine if it should pass through the surface or not.
Sort of a multisample implementation of alpha mixing for a path tracer.
I'm sure there are better ways of accomplishing this, but this one works nicely enough for me to commit it.