Fix improperly formatted thread count printout.
There was an edge case where the thread count was printed really weirdly
if the thread count was reduced due to a conflicting configuration
option. So if I rendered an image where the dimensions were like 50x50
pixels and we could only fit one tile in there, it would reduce the
thread count to 1 automatically, and then claim that we're rendering
with "-1+2" threads, which sort of makes sense, but looks too weird for
me to leave it like that.