Change textBuffer/lineBuffer semantics
Previously, if you encountered an "empty" token, you got back a pointer to a null byte.
Added checks for that, so you just receive a pure NULL you can check against.
This makes sense to me. At least at 5AM it does. I'll revert it tomorrow if it no longer makes sense.
It also didn't break any of the existing tests, so that's nice!