Start implementing a new BSDF style material system.
This is still rough around the edges, and will need a lot of work. But it's already looking pretty good so I'm committing it.
I ported the following materials to the new system: glass, metal, diffuse

Note: All of this worked on the first try, no bugs anywhere. Just smooth sailing!