Prune mix nodes while constructing.
If the two nodes passed in to a mix node are equal, there is no point
in constructing a mix node, so we just return one of the existing ones.
This does raise the issue that whatever is passed into B and factor will
already be constructed needlessly, but we can add a post-processing pass
later to clean up the graph.