Start work on a basic volume implementation
This mostly just sets up the supporting infrastructure, mainly that the
sampler is passed into the intersection finding logic to sample volumes
randomly, and also establishes instances as an abstract 'thing we can
hit' more than it was before.
Fairly small tweak to the instance API here, nobody even uses volumes
yet since the actual implementation is so buggy. But it's a start!