Fix cross-arch networking failure.
Turns out, the issue was likely the fact that size_t is 64 bits on my
system and 32 bits on the Power/ARM systems I was testing against, so
the two sides of the network protocol were assuming a different size for
the chunk header value. I'm now pinning it to be 64 bits wide on both
ends to resolve that ambiguity and it seems to work now. It's
conceivable that we will have scenes complex enough to require >4GB
transfers, so 64 bits makes sense for the header, which is just the
message size number.

Resolves #104