Add ability to set local thread count to 0 when network rendering
If valid network render clients were found, one can now set the local
render thread count to zero to offload all rendering tasks to remote
worker nodes.
This is useful for when I'm working on my laptop and want to do the
heavy lifting on the desktop system in my office.