Implement a simple progress bar for client asset/scene sync
I tested the robustness of the rendering protocol by attempting a
cluster render with my home desktop all the way from Pakistan. The
render worked fined, but actually sending the assets over a very weak
network link took a long time, with no indication that progress was
being made.
This resolves that. It now displays a progress bar for each client when
sending assets.