Rework scene loading to support interactive changes
Previously, the scene loader (just called loadJSON()) was built with the
assumption that it would always just be called once at the beginning of
a render, and it would parse and apply everything in one go.
This set of changes tries to make it more flexible, so preferences and
scene changes can be applied incrementally.
It's probably not quite there yet, and there is more work to do to
enable removing scene assets, for instance, but this is a good step in
the right direction.
The scene parsing/loading code has been an eyesore for a long time, so
these changes also make it slightly less ugly.

Related changes were made to the camera API (again) to hopefully make it
make more sense as well. It made absolutely no sense today when I looked
at it. I have no idea what I was thinking before.

The core idea with scene parsing now is that if we encounter invalid
input from JSON, we emit a warning and leave that value alone. If
incoming JSON doesn't contain something, we also skip it instead of
complaining and stopping there.