Reduce hitRecord struct again from 88 to 80 bytes
I made the observation that we never use the start of a lightRay for
shading, so I just use the incident ray direction instead.
I could have taken out surfaceNormal instead and computed it lazily from
the full lightRay, but the normal is used everywhere all the time for
shading, I think it's better to keep it precomputed in the struct.