Teach parseValueNode() to parse directly if no type is provided
It just passes the ball to parseTextureNode, which knows the different
cases we need. If it gets a string, it tries to load that file, and if
it gets an array, it'll interpret that as RGBA values. For the rest of
the cases, like a number, or for fresnel and map_range, we have the
exceptions in parseValueNode()