Refactor nextTile*() to return a pointer instead of a copy
While multiple threads do now read and write to the global tiles array,
the observation was made that any given tile is guaranteed to only be
written to by a single thread at a time. Thus, it's safe to just return
a pointer to that element of the array directly to simplify state
management somewhat.