Remove global vertex buffers
This was a really old vestige of the early days of c-ray in 2015 or so.
Removing it simplifies code and makes writing new mesh parsers easier.
This also makes implementing dynamic state much nicer, since there is no
longer a set of global buffers to shuffle.

Really happy with this change! :^)