Rewrite mesh parsing logic to be able to deal with multi-mesh files
Multiple interdependent changes in this single commit, unfortunately.
These include:
- Moving materials from meshes to instances
- Splitting mesh and instance parsing to two stages
- Removing support for the old mesh-global bsdf override

That last point might break some existing scenes, but the bsdf
"guessing" logic is fairly robust. The few scenes I cheched so far
pretty much just work, it mostly just grabs the colors and sets it to a
diffuse if it isn't sure what the material is supposed to look like,
which is a very safe default.

Moving materials to instances allows us to *finally* take full advantage
of the instancing system, allowing us to make multiple unique material
overrides to apply to each instance.

Not sure if splitting the instance/mesh parsing to two stages made the
code simpler. I did it because of the rather rigid array patterns we use
in c-ray which call realloc() for every insert, which broke instances
since those just use pointers the object they are instancing. Might do a
second cleanup pass once we've got some nice linked lists in here.