Rewrite file cache API to not hold global state
The cache is now passed around as a struct.
This file cache is used for network renders to store on-disk assets in
memory. The protocol could use a lot of work, and on-disk local caching
for clients would be nice so we don't send assets over and over again.
Also compression. Compression would be nice.