Add a cool rainbow sphere to hdr.json
Demonstrates the new combineHSL, mapRange and vecToValue nodes quite
well.

It extracts the uv, grabs the u component, remaps that 0.0-1.0 range to
0.0-360.0 and passes that in as the hue to the mapHSL node to get a nice
rainbow effect. That is then passed as the color to a metal bsdf to get
a reflective rainbow sphere, that is then tilted towards the camera to
show the pole.

Also silenced some unneeded warnings that came out of this. The sphere
parsing code is still old and needs to be updated to match mesh parsing
better.