Don't pass world to node constructors
Before, we were mostly passing the entire world struct just to access a
handful of things. This big ol' refactor changes it such that only
what is actually needed is passed in. The calling side gets a bit more
busy, but that mostly presents opportunities to rework the construction
code to fix that.

Many changes, should work the same as before. Tests pass (save for that
one I really need to fix)