Massively reduce calls to calloc() by ~92%
I ran heaptrack for the first time (excellent program!) and discovered
that my linebuffer code was doing massive amounts of tiny allocations
during the initial setup stage.
This change just allocates that stuff on the stack instead, like it
should be doing. Slightly uglier to set up a linebuffer now, but I might
wrap that stuff into a macro or something eventually.