Significantly reduce instance memory use (-5%)
Instead of deep-copying the entire material set for each and every
instance (144B * mtl_count), instead just set the resulting BSDF of
every material. These BSDF graphs are hash consed, and only contain the
relevant shading info/logic actually needed during the render.
The actual material struct is a vestige of the old material system, and
should be hidden deep into the bowels of the wavefront parsing code. The
instancing logic was literally just looking up the bsdf field in the
material anyway!

For a large scene with 5 million instances, I observed a 5% reduction in
peak memory usage.