Add sampler to vectorNode and implement vec_mix
vectocolor and vectovalue just pass NULL as the sampler input, which is
a horrible idea. But fixing that means adding redundant sampler inputs
to both valueNodes and colorNodes. So I'll do that a bit later.

Again, no new tests added. Just hacked up the current vectorNode tests
to construct and pass in a valid sampler for the tests.