Add sampler param to valueNodes and colorNodes too
Predicatably, my NULL hack didn't work at all. Kind of annoying to have
to pass this param to so many places, when it's only rarely used, but at
least everything is consistent now.
I guess an option would be to just set a ptr to the sampler in the
hitRecord and pass it around that way, but that feels wrong.