Remove transformType from transform struct
It was used in very few places, and was taking up an extra 4 bytes in
the transform struct, bloating it from a neat and even 128B to 132B. A
single transform now neatly fits in two cachelines.
Still wondering if this really is the most optimal way to store
transforms, or if there is some trick to save space here. We do use
transforms a lot!