Enable -Wdouble-promotion and fix all warnings
I recently read this excellent article that, among other things,
recommended I enable this warning. It's useful to know when floats are
being implicitly promoted, so I'm leaving this on.

This also caught a few seemingly accidental promotions in stb_image.h. I'm
using an old version, but I'll have to see if these should be
upstreamed.

Article link:
https://nullprogram.com/blog/2023/04/29/