Load SDL2 symbols at runtime
Instead of trying to link SDL2 portably at compile time, which requires
the help of CMake and leads to a bunch of ugly #ifdef gunk, we now look
for a shared library at runtime, and load only the symbols we actually
need.
A mostly gutted set of SDL headers is included, purely to provide a few
types for more ergonomic use of the SDL2 API. This bloats the source
tree by about ~160K unfortunately.