Simplify thread state monitoring & rename state members
I don't know why or how I came up with the most overbuilt and fragile
way of tracking the amount of active threads. I was carefully
manipulating a state variable that kept the current count, and then
another map array of threads that I had already checked to make sure
that count variable was accurate.
When really all I had to do was to loop over the threads, count the
inactive ones, and then check if that matches the total thread count.
Duh.

Also did some camelCase -> snake_case while I was in there. Should be a
separate commit, but I'm too busy for that.