Fix relative include path
This is one of those things I need to figure out one of these days. With
a CMake build system, you can put pretty much anything in between
quotes, and it'll find it even if the relative path is wrong. Then with
a makefile, they have to be mostly correct, and then finally on more
obscure platforms, the build blows up if they are even slightly
incorrect.
Huh!