Fix & simplify render cluster state updates
It was previously just broken, and overcomplicated. I think it was
trying to somehow aggregate performance metrics on the worker nodes, and
then the master node was supposed to work out performance figures from
that, but that was way more complicated than it need be.
Now we just sync the tile states, and let the existing performance
metrics/state monitoring logic handle the rest.
Also cleaned up some of the SDL UI code, as that was also overbuilt.

Starting a see a pattern here - Slightly younger me really liked to come
up with really complicated solutions to problems :^)