Add support for ccache
For CMake builds, it now checks if ccache is available on the system,
and enables it if found.
This isn't a big project, but it still provides a nice speed boost for
rebuilds, especially when switching between release and debug.

The way I have it structured is not optimal, I don't use a build
directory for the different build types, and instead just rebuild the
whole thing from scratch.
Should probably fix it somehow, I'm just so used to how it is now :D