Fix mtllib emitter parsing
I recently broke it while shuffling things around. The whole premise of
trying to come up with reasonable materials from the incredibly
restrictive mtllib description is a bit shaky at best, but we still
make a best effort to produce defaults, with the assumption that
materials get overridden with proper material graphs in json.