Implement very basic dynamic array
Until now, I tended to stay away from containers and other abstractions
- Slowing down to think of a simple way to implement something instead
of just throwing in a container was one rationale for that. But it's
pretty obvious now that not having even a basic dynamic array is just
making the code harder to read than it has to be, and it limits and
slows down development needlessly.
So here it is, the dumbest, most obvious dynamic array for C. Some
preprocessor magic makes it pretty slick to set it up for new types.
Sadly, structs do require typedefs before the macro can handle them, I
tried to find a way to deal with that, but didn't find anything.
3 basic tests are included in the new dyn_array test set.