Remove support for negative offsets in wavefront parser
I checked, and none of the meshes I've used for testing seem to
actually this.
Negative offsets would index off the end of vertex buffers, but
supporting this requires that we iterate over the wavefront file at
least twice - Once to count the totals, and a second pass to do the
actual parsing.
Seems like a very small gain for all that additional parsing required,
and I'm hoping to delete this parser entirely in the near future anyway,
so I don't really mind if it's not fully compliant.