Move more args handling to main()
Moving a lot of the state set up that is scattered in random places to
where that setup is actually needed.
Render plane now gets quantized right before starting the render,
instead of in the setup code.
We also now apply resolution overrides to a copy of the camera, instead
of overwriting the value for every camera in the actual scene state.