More rage-hacking, ripping out things and putting them in main()
Most overrides are now passed in from main, which is very good. Soon we
can rewrite the args parser to not have that awful global state.
main.c is getting awfully cluttered though, but that will be resolved
over time. We just have to trust the process.
I'm also moving pretty fast, so I'm likely adding lots of bugs, I'll
work on those once I've got the API and lib/driver separation to a
satisfactory state.