Remove global args from worker.c
Instead, pass port to worker client via API.
Now, the only remaining use of args_is_set() is in logr(). I think I'll
just have a global in there? Either way, code beyond the API boundary
shouldn't log to stdout, and should instead return errors or give us
data in a callback that we can print in the driver part, if we want to.
Not sure yet.