Actually make args table local to main()
It's now only available in that scope. Pretty good, though now I feel
like using the hashtable to store arguments like this is
overcomplicating it quite a bit. I think I might just replace it with a
little struct, might be quite a bit more ergonomic to read and use.