Implement T_arr_copy() & T_arr_join() + tests
Copy can be used to make a 'deep' copy of a dynamic array. It obviously
won't know about any memory that underlying elements may point to, but
doing a copy of the actual data may sometimes be useful.

Join appends array b to array a, and then frees array b. This is useful
when adding batches of multiple elements to an array.