Pass length to cJSON and improve status printout
Made human_file_size() slightly nicer, it can now take an optional
stack buffer to skip a dynamic allocation.
I observed that we already know the length of the JSON payload, and I
was just needlessly having cJSON do another strlen() on it, so this
saves a bit of time.