Start separating driver from library parts
Many things need to be done, but the aim is to have all the driver
program code live under src/driver/* - This should include all asset
loading and output code.

For now, there are still a lot of FIXMEs where internal components
depend on stuff under driver/, which needs to be sorted.