Rewrite hdr/background color parsing and handling
Instead of bespoke parsing code, we now (mostly) use the existing node
parsing logic, with only a bit of special sauce if the shader is the
background one.
I also moved the offset to be internal to the shader, and also now fully
node-enabled, so it can be controlled in the renderer node graph.