Implement a mesh + vertex data API
Not really that happy with it yet, but it's better than exposing
internal datatypes like before. Now, one builds a vertex buffer with
geometry data, and that can then be bound to one or more meshes.
Currently, all wavefront meshes in a single file share the same
geometry, so it's still refcounting the buffer internally.