Fix incorrect include path
This puzzled me for a while, for whatever reason, this path resolved to
be correct on my system. I spent an hour clearing caches and doing
research, until `cc -E` showed it was picking up a weird copy of
`utils/*` from one level outside the project root.
I have no idea why that directory was there. The files existed, but were
empty, and that directory was created in October 2022. It was picked up
because I have a `-I./includes` directive to add the main includes
directory, and that's how it was picked up.
Weird!