Start working on a Python wrapper
It's extremely non-pythonic at the moment, I still need to get into a
python state of mind. But this works!
Turns out, the qoi impl part was needed after all, so I put it back -
It's needed when linking against libc-ray.so as a standalone library.

This bare minimum python script links to it, sets up a few functions,
creates a renderer, and communicates with it a bit. Baby steps.