python: Fix indentation clobbered by Helix
By far the most frustrating part about Helix is the indentation
configuration. Not only do they default to spaces for all languages
(wtf?), they explicitly require the user to correct this option, by hand,
for each language they wish to use.
Not only that, even after fixing languages.toml, Helix still chooses to
ignore the setting there, if the file it managed to sneak spaces in
seems to mostly be indented with spaces.
I really think I should patch my own version of Helix to fix this mess,
it's really frustrating to have to fight tooling like this.