lib: Add a very thin wrapper around texture struct
It's not really even a wrapper, just a copy of the type. And the driver
does still use the internal texture type in places, since the lib and
driver both use that code.
I just wanted to get that type out of the public API so I can continue
with more python work.