Implement a CPython API wrapper
I decided to go this route instead, it's only ~600 lines of C, and a
good excuse to finally learn the CPython API a bit.

I didn't like blinly hooking into a .so with ctypes - It tends to be
quite fragile, as changes to the API silently break things. This way, I
at least give the Python interpreter an up to date view of the API.

To use this, invoke 'make pylib', and then you can 'import wrappers.cray
as cr' from the project root in the Python REPL.

It's pretty nifty! The API of course still needs tons of work, and this
is only the low-level shim to get things going. All of these functions
will be wrapped in convenience classes in Python land.

I changed cr_camera to be a cr_object instead of int32_t. Just seemed
like asking for trouble down the line for no reason to have that be
different.

For now though, this will do. It's getting a bit late. Quite happy with
the result! :^)