blender: Make the blender add-on sort of load properly
It's still very shaky, I managed to segfault the blender Python
interpreter quite a few times while figuring things out, so I'm enabling
faulthandler for the time being.

The segfault happens when I copy a fresh cray_wrap.so to the Blender
add-on directory, and then reload my add-on. Evidently, something
somewhere wasn't expecting that, and the interpreter segfaults on the
first call to cray_wrap, which happens to be in register, to get the
semantic version from libc-ray.