lib+blender: Fix up coordinate system mismatch
My brain has a hard time keeping up with all the different things going
on, but I got it to the point that the camera seems to work as expected.
I ended up going the easy route, and just adding a toggle to the API to
switch a specified camera to use Blender's native coordinate system, and
still defaults to c-ray coords (y up, +z forward) so existing json
scenes work normally.