blender: Fix Cycles shader nodes panel not showing up
We have to tell it that we want to use Cycles nodes instead of custom
ones.
The discoverability of the API is still super vague to me. I only
discovered this flag from a forum post - They are documented, but I have
to open web docs instead of just browsing code.