lib: Add targets to build & install blender add-on
make blsync syncs only the python scripts, which can then (most of the
time) be reloaded in Blender. make fullblsync also builds and swaps in
the library .so. Blender crashes if you do this while it's running and
then reload add-ons, so only use that when restarting it anyway.
I've also noticed that sometimes even scripts aren't reloaded properly,
so if you see buggy python errors, just fullblsync and restart Blender.