lib+blender: Various fixes to make material converting more robust
On the lib side, I detect and emit a warning if a shader buffer is empty
(which shouldn't happen, but can if the API is used 'incorrectly').
I also patch in a null material to make the renderer happy, since it
assumes that there is at least one material there.
On the blender side, I now convert the material trees once, instead of
repeating the conversion for every mesh, since meshes often share
materials.