lib: Make render result buffer presistent
We now only have a single float-precision buffer in the renderer, which
gets updated with cr_renderer_render(). Bit of a clearer API now.

With this change, I'm also finally removing the duplicate setPixel() +
sRGB conversion stuff from the core render threads, which was used for a
long time to facilitate smooth update of the render preview, at the a
cost to render performance.
Now, we only update the internal linear color data, and that data is
then converted on the main thread before passing it to SDL. Much better
design - I haven't measured how much this speeds up the actual
rendering, but removing that logic from the render threads is a good
change.