lib: Fix diverging prng sequences when enabling/disabling dof
Previously, we would only grab random values for the aperture
calculation if we were simulating an aperture.
This causes images to differ more than they need to, if the only
difference between them is the camera fstop setting.
Computing it every time, even if it isn't needed serves to maintain the
same rng sequence, which results in minimal differences between the two
resulting images.