lib+python: Fix up callbacks
I ended up going with the wrapper approach - c-ray receives a stub that
extracts Python arguments and calls the Python callback function while
holding the GIL.
It seems to work, I just hooked up a print to on_start. Will remain to
be seen if this approach is at all compatible with the way Blender does
things.