blender: Improve performance of display_bitmap()
Following a discussion on blender.chat, @Mysteryem pointed out that the
reason why I can't just pass a flat float array to foreach_set() in the
result rect may be a bug, so this may just be a temporary workaround.
Either way, this is way faster than what I had before, I also added
printouts to show how long it took.

I also experimented with trying to access the actual float array on the
C side, so I could just do a raw memcpy(), but I couldn't get that
working.