blender: Explicitly tear down/free c-ray & fix small leak
The destructor for the PyCapsule was never being called, despite me
deleting it in Python, so an explicit teardown function is now expected
to be called.
I don't know if it's specific to my system, but Blender seems to leak
quite a bit of memory, even while using Cycles. I fixed a small leak in
c-ray, but the rest seems to all be in Blender code.