lib+blender: Implement a very buggy and slow, but functional interactive mode :^)
It's not very pretty, and I'm doing slow copies of the bitmap on every
redraw, and only changes to a single camera get updated, but it works!
Lots of room for improvement, I spent quite a bit of time slowly
chipping away at this over the last few weeks, while doing other
improvements.