blender: Use a zero-copy buffer for interactive mode
I figured out a way to create a python buffer that just directly
references c-ray's bitmap buffer, so now I only need to create a new
texture at draw time, which seems moderately fast. 7-12ms was a common
range for the drawing logic, which is faster than it probably needs to
be, at least for now.