lib: Implement thread pool and use it for BVH building
I was intending on simplifying the main renderer thread management with
this as well, but on second thought, this isn't really a good fit for
that. The current approach gives us load balancing for free, since the
fixed render threads just pick up new tiles to render as soon as they
finish the previous one.
Using this thread pool for rendering will require a bit more planning.