lib: Improve thread_start() and delete thread_user_data()
Since the struct has been public for a while now, the thread_user_data()
abstraction to get at that pointer is no longer needed.
We also extend the lifetime of struct cr_thread to enable callers of
thread_start() to pass in the address of a stack-local cr_thread struct,
if they choose to do that.
No more need for the static hack in the interactive renderer wrapper
function.