blender: Handle mesh transformations interactively
One can now rotate, scale and move around things in shaded viewport
mode, and the changes are reflected in real-time.
Caveat: If you scale an object too small, c-ray hits a fatal error
trying to find an inverse for that transform, and brings Blender down
with a call to abort() XD

TODO: Make it not do that.