lib: Improve test & fix race condition in thread pool
It could sometimes happen that we enqueue a small amount of work, and
then manage to lock the pool mutex in a subsequent call to
thread_pool_wait() before any threads managed to grab any work. Since we
didn't check if the pool has any queued work, it could happen that we
continue executing on the main thread assuming the pool is done working,
when in reality it actually just started working.
Simple fix, just check if there is work to be done, and block as long as
we do.