lib: Un-GNU-ify C codebase
Until now, I had specified -std=c99, but defined _GNU_SOURCE, meaning I
was using a bunch of GNU extensions. It's comfy to write code that way,
but portability is even more comfy. I enabled -Wpedantic for this pass,
but I just can't bear cluttering the dlsym() stuff with unnecessary
casts just because ia64 was weird back in the day, so I didn't leave it
on. I'll check on things now and then, but for now, this is a good start
to better POSIX C99 adherence :]