lib: Use custom extension type for renderer callback info
Apparently the best practice for CPython plugins is to define the needed
types on the C side rather than wrapping the plugin with a bunch of
extra ctypes gunk to wrap the underlying structs.
This seems to work, it's even more boilerplate for every type, but we
get nice python types out on the other end, so it's probably worth it.