lib: Use rwlock + copies in BVH updates
Now that we're building BVHs for meshes while simultaneously doing other
preparations, we can't really refer directly to the scene mesh array in
the async builder, since that array may end up getting realloc'd during
a BVH build. Now, we grab a copy of the mesh struct and the index, which
we can then later use to update the actual mesh, guarded by a rwlock.