lib: Decode textures asynchronously
Rename bvh_builder -> bg_worker, and now queue texture decode tasks on
there as well.
Not fully happy with the way failed decodes are handled, we just end up
with a blank texture that then has to be checked in the texture node
while rendering, which is probably not ideal. Before, a failed texture
decode just passed NULL to newImageTexture(), which then checked that
and returned NULL. All nodes that take a colorNode input have sane
defaults for that case, but we can't really do it that way with this
approach.