blender: Fix bug where c-ray was never deleted
Turns out, I was passing self into the callback args, which presumably
then held a reference to it, which meant that the Blender python runtime
never calls __del__() on it. I wasn't even using self for anything, I
was just grabbing the current sample count from it.
Fix is to just pass in the sample count directly, instead of passing in
self.