lib: Start porting c-ray python types to native ones
Initially I just slapped in a bunch of ctypes gunk because that seemed
like the most obvious approach, but I've since discovered the 'proper'
way to expose custom Python types, which this commit now does for
cr_bitmap and cr_renderer_cb_info.
The end goal is to no longer call the CPython API from the python
wrapper for any reason, and to get rid of ctypes boilerplate.