build: Check for ccache in Makefile
The CMake build system exists primarily for Windows support, which is on
its way out, because WSL2 is quite mature now.
I've been mainly focused on the GNU Makefile build system lately, so I'm
porting some nice things the CMake build enabled there now.
It's fairly easy to check and enable this in GNU Make, so it now checks
USE_CCACHE to see if ccache should be used (defaults to 1), then it
checks if ccache is available, and if so, prepends `ccache ` to CC.