lib: Make resizing viewport in interactive mode work
It's barely functional. The whole heap of overly complicated state
management for the interactive rendering mode is a giant mess, and I
want to redo the whole thing. But this works for now. It now recomputes
camera parameters and tiles when resizing, and does the right dance with
state flags and mutices to keep the renderer from blowing up. There are
likely edge cases where that still happens, but resizing the viewport
now works when using shaded viewport mode in Blender!