lib+blender: Handle adding objects while in interactive mode
Found some flaws in the BVH update logic, as well as a missing
sync_mesh() call on the python side.
I actually do need to completely rewrite the Blender depsgraph syncing
logic from the ground up. The way it's structured right now doesn't
really match how the depsgraph and partial updates are supposed to be
handled.