driver: Don't block on window init in renderer on_start() callback
I can't find documentation on this, but I've observed an annoying
behaviour where SDL_CreateWindow() will block infinitely if c-ray
is started from a terminal that is full-screen on a tiling window
manager, such as i3. This then needlessly delays rendering until
either the terminal exits fullscreen, or the cursor is moved to
another display. Weird behaviour, but I can't be bothered to look
into it much further.
We now do the slow window init in a background thread, which has
the added benefit of speeding up boot-up time by around ~250ms on
my system.