driver: Use full soname for libSDL2, not Debian-ism / CMake build-ism
The upstream autotools-based SDL2 build system does not install a
libSDL2-2.0.so, it installs libSDL2-2.0.so.0 as the actual library, and
libSDL2.so for the symlink used at link time with -l (but not intended
to be required at run time). Debian also installs an extra
libSDL2-2.0.so symlink in the libsdl2-dev package, which is what this
was using, as does upstream's newer CMake-based build system. Older
FreeBSD ports versions used autotools and so do not have such a symlink.
Moreover, we don't want to require the development package to be
installed either, just the actual library implementation, so even if
this were reliably available, we shouldn't be using it anyway.

Note that since we vendor the headers used to interface with the
library, hard-coding the soname is fine; we won't be ABI-compatible with
other sonames.