SDL2: Fix SDL_BYTEORDER under project's build flags
FreeBSD's sys/endian.h gates the definition of the non-underscored
BYTE_ORDER with __BSD_VISIBLE || _POSIX_C_SOURCE > 200809. That is, the
default compiler options (or -D_BSD_SOURCE) will cause it to be defined,
but due to -D_POSIX_C_SOURCE=200112L being added to the compiler flags,
neither check will hold, and so it won't be defined, leading to it being
effectively 0 when checked. On NetBSD this define is instead gated on
defined(_XOPEN_SOURCE) || defined(_NETBSD_SOURCE), and I don't know
whether it ends up defining those somehow, but the underscored define is
always available, so this should work there.