lib: Shorten node insert/reuse log messages
These messages are quite noise even for simple node graphs, so we only
emit them in the 'spam' log level. It makes sense to then ensure we
don't print redundant info, which we do here. We now only print the node
dump, where green indicates a hashtable hit, and red indicates a miss.