lib+driver+common+tests: Introduce initial version of v.h
I'm building a collection of useful datastructures and other widgets
I find myself copying to other projects/experiments. This collection
will eventually all live in a new single-header library with the working
name of `v.h`. It will eventually live in a separate repo if it becomes
useful enough.

For the time being, it'll live in c-ray's `include/` while I migrate
things to it. This initial version already has the following components:

- v_mod: Runtime library loading abstraction (posix + win32)
- v_timer: Small timer gadget
- v_arr: Dynamic arrays

The corresponding implementations in c-ray have been removed, and uses
have been updated to use the ones in `v.h`.