v.h: Make v_thread API more robust and nicer to use
The old approach was mimicking the pthreads API, but my aim with v.h is
to build wrappers for things where I find myself double-checking the
semantics from manpages every time I use them.

The new v_thread API is thus a much simplified create/destroy interface,
and it looks like this:

v_thread *v_thread_create(v_thread_ctx c, enum v_thread_type type);
void *v_thread_wait_and_destroy(v_thread *);

v_thread_create() will return an opaque pointer to a v_thread when the
type == v_thread_type_joinable. This opaque pointer must then be passed
to v_thread_wait_and_destroy(), which, as the name suggests, waits for
that thread to finish and then frees up the allocation associated with
it.
When creating a detached thread by passing v_thread_type_detached in the
create call, the return value must be used strictly for error checking.
NULL means something went wrong, non-NULL means the detached thread was
successfully created.
Because the v_thread allocation is automatically freed by the internal
thread stub when type == v_thread_type_detached, it's considered
undefined in this API to pass that pointer to
v_thread_wait_and_destroy(), since that may or may not result in
use-after-free. Waiting on a detached thread doesn't make sense either
way, so the users of this API are expected to know not to do that. A
note is included as a reminder about this.