v.h: Improve v_timer API
Changed signature from

void v_timer_start(v_timer *t);

into

v_timer v_timer_start(void);

Which reflects the usage of it in c-ray better. The intent with the
original one was to be able to restart the same timer, but this reads
nicer and supports that use case anyway.