driver: Remove unusual display.isBorderless option
I added it very early on, but never used it. Kind of a weird random
option to have.

This whole JSON parsing + standalone executable stuff is feeling more
and more deprecated. I only really play around with c-ray in Blender
now, so the only useful use-case the standalone bin/c-ray is to run it
as a worker node for clustered rendering. But that feature relies on the
fragile manually written JSON [de]serialization gunk to work, and I'd
really like to nuke all that stuff to focus on the renderer itself.
Maybe the protocol should wrap the more carefully considered public API
now that that exists? i.e. forwarding input from that to worker nodes,
instead of syncing all internal state after the fact?

The active maintenance required by the JSON stuff makes me want to work
on c-ray less. 70439b4 shows the amount of boilerplate required to add a
single node, which is too much.